src/requirements.txt
pytest==8.3.2
pytest-cov==5.0.0
pyyaml==6.0.1
